Checkpoint

+++ Contractor/Unfinished/Share.cs
@@ -4,9 +4,9 @@
 
     public class Share
     {
-        public virtual short Id { get; protected set; }
+        public virtual ushort Id { get; protected set; }
         public virtual Person Holder { get; set; }
-        public virtual int Quantity { get; set; }
+        public virtual ushort Quantity { get; set; }
         public virtual DateTime HeldFrom { get; set; }
         public virtual DateTime HeldTo { get; set; }
         public virtual DateTime LastUpdated { get; protected set; }
